+++ packages/client/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- **Utils**:
+  - `mergeUint8` now takes an Array<Uint8Array>. Soft deprecate old overload.
+
 ## 1.12.2 - 2025-05-27
 
 ### Fixed